Streamlit_Agentic_AI/requirements.txt
# requirements.txt
streamlit
pytesseract
pdf2image
opencv-python-headless
Pillow
scikit-learn
python-dateutil
joblib
numpy
pandas
